+++ src/command_modules/azure-cli-appservice/azure/cli/command_modules/appservice/custom.py
@@ -53,7 +53,8 @@ from ._appservice_utils import _generic_site_operation
 from ._create_util import (zip_contents_from_dir, get_runtime_version_details, create_resource_group,
                            should_create_new_rg, set_location, should_create_new_asp, should_create_new_app,
                            get_lang_from_content, get_num_apps_in_asp)
-from ._constants import (NODE_RUNTIME_NAME, OS_DEFAULT, STATIC_RUNTIME_NAME, PYTHON_RUNTIME_NAME, RUNTIME_TO_IMAGE)
+from ._constants import (NODE_RUNTIME_NAME, OS_DEFAULT, STATIC_RUNTIME_NAME, PYTHON_RUNTIME_NAME,
+                         RUNTIME_TO_IMAGE, NODE_VERSION_DEFAULT)
 
 logger = get_logger(__name__)
 
@@ -78,7 +79,7 @@ def create_webapp(cmd, resource_group_name, name, plan, runtime=None, startup_fi
     if not plan_info:
         raise CLIError("The plan '{}' doesn't exist".format(plan))
     is_linux = plan_info.reserved
-    node_default_version = '8.11.1'
+    node_default_version = NODE_VERSION_DEFAULT
     location = plan_info.location
     site_config = SiteConfig(app_settings=[])
     webapp_def = Site(location=location, site_config=site_config, server_farm_id=plan_info.id, tags=tags)
@@ -288,7 +289,8 @@ def enable_zip_deploy(cmd, resource_group_name, name, src, timeout=None, slot=No
         logger.warning("Starting zip deployment")
         requests.post(zip_url, data=zip_content, headers=headers)
     # check the status of async deployment
-    response = _check_zip_deployment_status(deployment_status_url, authorization, timeout)
+    response = _check_zip_deployment_status(cmd, resource_group_name, name, deployment_status_url,
+                                            authorization, timeout)
     return response
 
 
@@ -2077,7 +2079,7 @@ def create_function(cmd, resource_group_name, name, storage_account, plan=None,
     # adding appsetting to site to make it a function
     site_config.app_settings.append(NameValuePair(name='AzureWebJobsStorage', value=con_string))
     site_config.app_settings.append(NameValuePair(name='AzureWebJobsDashboard', value=con_string))
-    site_config.app_settings.append(NameValuePair(name='WEBSITE_NODE_DEFAULT_VERSION', value='8.11.1'))
+    site_config.app_settings.append(NameValuePair(name='WEBSITE_NODE_DEFAULT_VERSION', value='10.14.1'))
 
     if consumption_plan_location is None and not is_plan_elastic_premium(plan_info):
         site_config.always_on = True
@@ -2176,25 +2178,29 @@ def list_locations(cmd, sku, linux_workers_enabled=None):
     return client.list_geo_regions(full_sku, linux_workers_enabled)
 
 
-def _check_zip_deployment_status(deployment_status_url, authorization, timeout=None):
+def _check_zip_deployment_status(cmd, rg_name, name, deployment_status_url, authorization, timeout=None):
     import requests
     total_trials = (int(timeout) // 2) if timeout else 450
     num_trials = 0
     while num_trials < total_trials:
         time.sleep(2)
         response = requests.get(deployment_status_url, headers=authorization)
+        time.sleep(2)
         res_dict = response.json()
         num_trials = num_trials + 1
         if res_dict.get('status', 0) == 3:
-            raise CLIError("Zip deployment failed. {}".format(res_dict))
+            _configure_default_logging(cmd, rg_name, name)
+            raise CLIError("""Zip deployment failed. {}. Please run the command az webapp log tail
+                           -n {} -g {}""".format(res_dict, name, rg_name))
         elif res_dict.get('status', 0) == 4:
             break
         if 'progress' in res_dict:
             logger.info(res_dict['progress'])  # show only in debug mode, customers seem to find this confusing
     # if the deployment is taking longer than expected
     if res_dict.get('status', 0) != 4:
-        raise CLIError("""Deployment is taking longer than expected. Please verify
-                            status at '{}' beforing launching the app""".format(deployment_status_url))
+        _configure_default_logging(cmd, rg_name, name)
+        raise CLIError("""Timeout reached by the command, however, the deployment operation
+                       is still on-going. Navigate to your scm site to check the deployment status""")
     return res_dict
 
 
@@ -2254,11 +2260,14 @@ def get_history_triggered_webjob(cmd, resource_group_name, name, webjob_name, sl
     return client.web_apps.list_triggered_web_job_history(resource_group_name, name, webjob_name)
 
 
-def create_deploy_webapp(cmd, name, location=None, sku=None, dryrun=False):  # pylint: disable=too-many-statements
+def webapp_up(cmd, name, location=None, sku=None, dryrun=False, logs=False):  # pylint: disable=too-many-statements, too-many-branches
     import os
+    from azure.cli.core._profile import Profile
     client = web_client_factory(cmd.cli_ctx)
     # the code to deploy is expected to be the current directory the command is running from
     src_dir = os.getcwd()
+    user = Profile().get_current_account_user()
+    user = user.split('@', 1)[0]
     # if dir is empty, show a message in dry run
     do_deployment = False if os.listdir(src_dir) == [] else True
     _create_new_rg = True
@@ -2297,8 +2306,8 @@ def create_deploy_webapp(cmd, name, location=None, sku=None, dryrun=False):  # p
     location = set_location(cmd, sku, location)
     loc_name = location.replace(" ", "").lower()
     is_linux = True if os_val == 'Linux' else False
-    asp = "appsvc_asp_{}_{}".format(os_val, loc_name)
-    rg_name = "appsvc_rg_{}_{}".format(os_val, loc_name)
+    asp = "{}_asp_{}_{}".format(user, os_val, loc_name)
+    rg_name = "{}_asp_{}_{}".format(user, os_val, loc_name)
     # Resource group: check if default RG is set
     default_rg = cmd.cli_ctx.config.get('defaults', 'group', fallback=None)
     _create_new_rg = should_create_new_rg(cmd, default_rg, rg_name, is_linux)
@@ -2348,12 +2357,15 @@ def create_deploy_webapp(cmd, name, location=None, sku=None, dryrun=False):  # p
         logger.warning("App service plan '%s' already exists.", asp)
         _show_too_many_apps_warn = get_num_apps_in_asp(cmd, rg_name, asp) > 5
         _create_new_app = should_create_new_app(cmd, rg_name, name)
+
     # create the app
     if _create_new_app:
-        logger.warning("Creating app '%s' ....", name)
+        logger.warning("Creating app '%s' ...", name)
         create_webapp(cmd, rg_name, name, asp, runtime_version if is_linux else None)
         logger.warning("Webapp creation complete")
         _set_build_app_setting = True
+        # Configure default logging
+        _configure_default_logging(cmd, rg_name, name)
         if _show_too_many_apps_warn:
             logger.warning("There are sites that have been deployed to the same hosting "
                            "VM of this region, to prevent performance impact please "
@@ -2373,17 +2385,17 @@ def create_deploy_webapp(cmd, name, location=None, sku=None, dryrun=False):  # p
                            site_config.windows_fx_version, runtime_version)
             update_site_configs(cmd, rg_name, name, windows_fx_version=runtime_version)
 
-        if do_deployment and not is_skip_build:
-            _set_build_app_setting = True
-            # setting the appsettings causes a app restart so we avoid if not needed
-            application_settings = client.web_apps.list_application_settings(rg_name, name)
-            _app_settings = application_settings.properties
-            for key, value in _app_settings.items():
-                if key.upper() == 'SCM_DO_BUILD_DURING_DEPLOYMENT' and value.upper() == "FALSE":
-                    _set_build_app_setting = False
+    if do_deployment and not is_skip_build:
+        _set_build_app_setting = True
+        # setting the appsettings causes a app restart so we avoid if not needed
+        application_settings = client.web_apps.list_application_settings(rg_name, name)
+        _app_settings = application_settings.properties
+        for key, value in _app_settings.items():
+            if key.upper() == 'SCM_DO_BUILD_DURING_DEPLOYMENT' and value.upper() == "FALSE":
+                _set_build_app_setting = False
+                break
 
     # update create_json to include the app_url
-    url = _get_url(cmd, rg_name, name)
     if _set_build_app_setting:
         # setting to build after deployment
         logger.warning("Updating app settings to enable build after deployment")
@@ -2403,13 +2415,14 @@ def create_deploy_webapp(cmd, name, location=None, sku=None, dryrun=False):  # p
             os.remove(zip_file_path)
         except OSError:
             pass
-    create_json.update({'app_url': url})
-    logger.warning("All done.")
-    return create_json
+    if logs:
+        _configure_default_logging(cmd, rg_name, name)
+    logger.warning("All done. Launching the app in your default browser.")
+    return view_in_browser(cmd, rg_name, name, None, logs)
 
 
 def _ping_scm_site(cmd, resource_group, name):
-    #  wakeup kudu, by making an SCM call
+    #  wake up kudu, by making an SCM call
     import requests
     #  work around until the timeout limits issue for linux is investigated & fixed
     user_name, password = _get_site_credential(cmd.cli_ctx, resource_group, name)
@@ -2523,3 +2536,10 @@ def create_devops_build(cmd, functionapp_name=None, organization_name=None, proj
                                                                  organization_name, project_name, repository_name,
                                                                  overwrite_yaml, allow_force_push, use_local_settings)
     return azure_devops_build_interactive.interactive_azure_devops_build()
+
+
+def _configure_default_logging(cmd, rg_name, name):
+    logger.warning("Configuring default logging for the app, if not already enabled")
+    return config_diagnostics(cmd, rg_name, name,
+                              application_logging=True, web_server_logging='filesystem',
+                              docker_container_logging='true')