+++ src/command_modules/azure-cli-appservice/HISTORY.rst
@@ -2,7 +2,9 @@
 
 Release History
 ===============
-
+* functionapp: add feature for establishing CI CD to an Azure DevOps pipeline from a Github repository
+* functionapp: in `az functionapp devops-build create`, added `--github-pat` flag to accept Github personal access token
+* functionapp: in `az functionapp devops-build create`, added `--github-repository` flag to accept Github repository that contains a functionapp source code
 * webapp: az webapp up --logs was failing with a error and updating default .NETCORE version to 2.1
 
 0.2.17