+++ src/command_modules/azure-cli-appservice/HISTORY.rst
@@ -3,8 +3,11 @@
 Release History
 ===============
 
+* webapp: az webapp up --logs was failing with a error and updating default .NETCORE version to 2.1
+
 0.2.17
 ++++++
+* webapp: Adding --logs support to az webapp up and other improvements to the up command
 * functionapp: fix `az functionapp devops-build create` command azure-pipelines.yml generation issues
 * functionapp: improve `az functionapp devops-build create` error handlings and error indications
 * functionapp: remove `--local_git` flag in devops-build command, local git detection and handling are compulsory for creating Azure DevOps pipelines